Use a unique temporary directory for AnalyzerAssemblyLoader
When multiple people try to do .NET development on a single system, the
first one to create /tmp/VBCSCompiler succeeds, and the others get
cryptic errors like this:

    Unable to load Analyzer assembly runtime/.dotnet/sdk/7.0.100/Sdks/Microsoft.NET.Sdk/analyzers/ILLink.RoslynAnalyzer.dll :
    Could not load file or assembly 'ILLink.RoslynAnalyzer, Version=7.0.6.47103, Culture=neutral, PublicKeyToken=31bf3856ad364e35'.
    Access is denied.

The root cause is the inability of
ShadowCopyAnalyzerAssemblyLoader.CreateUniqueDirectoryForProcess() to
create /tmp/VBCSCompiler, which is already owned by a different user.

Fix by adding a random suffix.